+++ examples/all-in-one.py
@@ -1,9 +1,11 @@
 #!/usr/bin/env python3
 
-import time
 import colorsys
 import sys
-import ST7735
+import time
+
+import st7735
+
 try:
     # Transitional fix for breaking change in LTR559
     from ltr559 import LTR559
@@ -11,19 +13,20 @@ try:
 except ImportError:
     import ltr559
 
-from bme280 import BME280
-from pms5003 import PMS5003, ReadTimeoutError as pmsReadTimeoutError
-from enviroplus import gas
-from PIL import Image
-from PIL import ImageDraw
-from PIL import ImageFont
-from fonts.ttf import RobotoMedium as UserFont
 import logging
 
+from bme280 import BME280
+from fonts.ttf import RobotoMedium as UserFont
+from PIL import Image, ImageDraw, ImageFont
+from pms5003 import PMS5003
+from pms5003 import ReadTimeoutError as pmsReadTimeoutError
+
+from enviroplus import gas
+
 logging.basicConfig(
-    format='%(asctime)s.%(msecs)03d %(levelname)-8s %(message)s',
+    format="%(asctime)s.%(msecs)03d %(levelname)-8s %(message)s",
     level=logging.INFO,
-    datefmt='%Y-%m-%d %H:%M:%S')
+    datefmt="%Y-%m-%d %H:%M:%S")
 
 logging.info("""all-in-one.py - Displays readings from all of Enviro plus' sensors
 
@@ -38,11 +41,11 @@ bme280 = BME280()
 pms5003 = PMS5003()
 
 # Create ST7735 LCD display class
-st7735 = ST7735.ST7735(
+st7735 = st7735.ST7735(
     port=0,
     cs=1,
-    dc=9,
-    backlight=12,
+    dc="GPIO9",
+    backlight="GPIO12",
     rotation=270,
     spi_speed_hz=10000000
 )
@@ -54,7 +57,7 @@ WIDTH = st7735.width
 HEIGHT = st7735.height
 
 # Set up canvas and font
-img = Image.new('RGB', (WIDTH, HEIGHT), color=(0, 0, 0))
+img = Image.new("RGB", (WIDTH, HEIGHT), color=(0, 0, 0))
 draw = ImageDraw.Draw(img)
 font_size = 20
 font = ImageFont.truetype(UserFont, font_size)
@@ -74,7 +77,7 @@ def display_text(variable, data, unit):
     vmax = max(values[variable])
     colours = [(v - vmin + 1) / (vmax - vmin + 1) for v in values[variable]]
     # Format the variable name and value
-    message = "{}: {:.1f} {}".format(variable[:4], data, unit)
+    message = f"{variable[:4]}: {data:.1f} {unit}"
     logging.info(message)
     draw.rectangle((0, 0, WIDTH, HEIGHT), (255, 255, 255))
     for i in range(len(colours)):
@@ -141,7 +144,7 @@ try:
         # One mode for each variable
         if mode == 0:
             # variable = "temperature"
-            unit = "C"
+            unit = "°C"
             cpu_temp = get_cpu_temperature()
             # Smooth out with some averaging to decrease jitter
             cpu_temps = cpu_temps[1:] + [cpu_temp]