+++ examples/light.py
@@ -1,7 +1,8 @@
 #!/usr/bin/env python3
 
-import time
 import logging
+import time
+
 try:
     # Transitional fix for breaking change in LTR559
     from ltr559 import LTR559
@@ -11,9 +12,9 @@ except ImportError:
 
 
 logging.basicConfig(
-    format='%(asctime)s.%(msecs)03d %(levelname)-8s %(message)s',
+    format="%(asctime)s.%(msecs)03d %(levelname)-8s %(message)s",
     level=logging.INFO,
-    datefmt='%Y-%m-%d %H:%M:%S')
+    datefmt="%Y-%m-%d %H:%M:%S")
 
 logging.info("""light.py - Print readings from the LTR559 Light & Proximity sensor.
 
@@ -25,9 +26,9 @@ try:
     while True:
         lux = ltr559.get_lux()
         prox = ltr559.get_proximity()
-        logging.info("""Light: {:05.02f} Lux
-Proximity: {:05.02f}
-""".format(lux, prox))
+        logging.info(f"""Light: {lux:05.02f} Lux
+Proximity: {prox:05.02f}
+""")
         time.sleep(1.0)
 except KeyboardInterrupt:
     pass