+++ examples/sensorcommunity.py
@@ -1,38 +1,35 @@
 #!/usr/bin/env python3
 
-import requests
-import ST7735
-import time
-from bme280 import BME280
-from pms5003 import PMS5003, ReadTimeoutError, ChecksumMismatchError
-from subprocess import PIPE, Popen, check_output
-from PIL import Image, ImageDraw, ImageFont
-from fonts.ttf import RobotoMedium as UserFont
-
-try:
-    from smbus2 import SMBus
-except ImportError:
-    from smbus import SMBus
 import logging
+import time
+from subprocess import check_output
+
+import requests
+import st7735
+from bme280 import BME280
+from fonts.ttf import RobotoMedium as UserFont
+from PIL import Image, ImageDraw, ImageFont
+from pms5003 import PMS5003, ChecksumMismatchError, ReadTimeoutError
+from smbus2 import SMBus
 
 logging.basicConfig(
-    format='%(asctime)s.%(msecs)03d %(levelname)-8s %(message)s',
+    format="%(asctime)s.%(msecs)03d %(levelname)-8s %(message)s",
     level=logging.INFO,
-    datefmt='%Y-%m-%d %H:%M:%S')
+    datefmt="%Y-%m-%d %H:%M:%S")
 
-logging.info("""luftdaten.py - Reads temperature, pressure, humidity,
-#PM2.5, and PM10 from Enviro plus and sends data to Luftdaten,
-#the citizen science air quality project.
+logging.info("""sensorcommunity.py - Reads temperature, pressure, humidity,
+PM2.5, and PM10 from Enviro plus and sends data to Sensor.Community,
+the citizen science air quality project.
 
-#Note: you'll need to register with Luftdaten at:
-#https://meine.luftdaten.info/ and enter your Raspberry Pi
-#serial number that's displayed on the Enviro plus LCD along
-#with the other details before the data appears on the
-#Luftdaten map.
+Note: you'll need to register with Sensor.Community at:
+https://devices.sensor.community/ and enter your Raspberry Pi
+serial number that's displayed on the Enviro plus LCD along
+with the other details before the data appears on the
+Sensor.Community map.
 
-#Press Ctrl+C to exit!
+Press Ctrl+C to exit!
 
-#""")
+""")
 
 bus = SMBus(1)
 
@@ -40,11 +37,11 @@ bus = SMBus(1)
 bme280 = BME280(i2c_dev=bus)
 
 # Create LCD instance
-disp = ST7735.ST7735(
+disp = st7735.ST7735(
     port=0,
     cs=1,
-    dc=9,
-    backlight=12,
+    dc="GPIO9",
+    backlight="GPIO12",
     rotation=270,
     spi_speed_hz=10000000
 )
@@ -62,15 +59,15 @@ def read_values():
     cpu_temp = get_cpu_temperature()
     raw_temp = bme280.get_temperature()
     comp_temp = raw_temp - ((cpu_temp - raw_temp) / comp_factor)
-    values["temperature"] = "{:.2f}".format(comp_temp)
-    values["pressure"] = "{:.2f}".format(bme280.get_pressure() * 100)
-    values["humidity"] = "{:.2f}".format(bme280.get_humidity())
+    values["temperature"] = f"{comp_temp:.2f}"
+    values["pressure"] = f"{bme280.get_pressure() * 100:.2f}"
+    values["humidity"] = f"{bme280.get_humidity():.2f}"
     try:
         pm_values = pms5003.read()
         values["P2"] = str(pm_values.pm_ug_per_m3(2.5))
         values["P1"] = str(pm_values.pm_ug_per_m3(10))
     except(ReadTimeoutError, ChecksumMismatchError):
-        logging.info("Failed to read PMS5003. Reseting and retrying.")
+        logging.info("Failed to read PMS5003. Resetting and retrying.")
         pms5003.reset()
         pm_values = pms5003.read()
         values["P2"] = str(pm_values.pm_ug_per_m3(2.5))
@@ -88,15 +85,15 @@ def get_cpu_temperature():
 
 # Get Raspberry Pi serial number to use as ID
 def get_serial_number():
-    with open('/proc/cpuinfo', 'r') as f:
+    with open("/proc/cpuinfo", "r") as f:
         for line in f:
-            if line[0:6] == 'Serial':
+            if line.startswith("Serial"):
                 return line.split(":")[1].strip()
 
 
 # Check for Wi-Fi connection
 def check_wifi():
-    if check_output(['hostname', '-I']):
+    if check_output(["hostname", "-I"]):
         return True
     else:
         return False
@@ -108,10 +105,12 @@ def display_status():
     text_colour = (255, 255, 255)
     back_colour = (0, 170, 170) if check_wifi() else (85, 15, 15)
     id = get_serial_number()
-    message = "{}\nWi-Fi: {}".format(id, wifi_status)
-    img = Image.new('RGB', (WIDTH, HEIGHT), color=(0, 0, 0))
+    message = f"{id}\nWi-Fi: {wifi_status}"
+    img = Image.new("RGB", (WIDTH, HEIGHT), color=(0, 0, 0))
     draw = ImageDraw.Draw(img)
-    size_x, size_y = draw.textsize(message, font)
+    x1, y1, x2, y2 = font.getbbox(message)
+    size_x = x2 - x1
+    size_y = y2 - y1
     x = (WIDTH - size_x) / 2
     y = (HEIGHT / 2) - (size_y / 2)
     draw.rectangle((0, 0, 160, 80), back_colour)
@@ -119,7 +118,7 @@ def display_status():
     disp.display(img)
 
 
-def send_to_luftdaten(values, id):
+def send_to_sensorcommunity(values, id):
     pm_values = dict(i for i in values.items() if i[0].startswith("P"))
     temp_values = dict(i for i in values.items() if not i[0].startswith("P"))
 
@@ -133,7 +132,7 @@ def send_to_luftdaten(values, id):
         resp_pm = requests.post(
             "https://api.sensor.community/v1/push-sensor-data/",
             json={
-                "software_version": "enviro-plus 0.0.1",
+                "software_version": "enviro-plus 1.0.0",
                 "sensordatavalues": pm_values_json
             },
             headers={
@@ -145,17 +144,17 @@ def send_to_luftdaten(values, id):
             timeout=5
         )
     except requests.exceptions.ConnectionError as e:
-        logging.warning('Sensor.Community (Luftdaten) PM Connection Error: {}'.format(e))
+        logging.warning(f"Sensor.Community PM Connection Error: {e}")
     except requests.exceptions.Timeout as e:
-        logging.warning('Sensor.Community (Luftdaten) PM Timeout Error: {}'.format(e))
+        logging.warning(f"Sensor.Community PM Timeout Error: {e}")
     except requests.exceptions.RequestException as e:
-        logging.warning('Sensor.Community (Luftdaten) PM Request Error: {}'.format(e))
+        logging.warning(f"Sensor.Community PM Request Error: {e}")
 
     try:
         resp_bmp = requests.post(
             "https://api.sensor.community/v1/push-sensor-data/",
             json={
-                "software_version": "enviro-plus 0.0.1",
+                "software_version": "enviro-plus 1.0.0",
                 "sensordatavalues": temp_values_json
             },
             headers={
@@ -167,17 +166,17 @@ def send_to_luftdaten(values, id):
             timeout=5
         )
     except requests.exceptions.ConnectionError as e:
-        logging.warning('Sensor.Community (Luftdaten) Climate Connection Error: {}'.format(e))
+        logging.warning(f"Sensor.Community Climate Connection Error: {e}")
     except requests.exceptions.Timeout as e:
-        logging.warning('Sensor.Community (Luftdaten) Climate Timeout Error: {}'.format(e))
+        logging.warning(f"Sensor.Community Climate Timeout Error: {e}")
     except requests.exceptions.RequestException as e:
-        logging.warning('Sensor.Community (Luftdaten) Climate Request Error: {}'.format(e))
+        logging.warning(f"Sensor.Community Climate Request Error: {e}")
 
     if resp_pm is not None and resp_bmp is not None:
         if resp_pm.ok and resp_bmp.ok:
             return True
         else:
-            logging.warning('Luftdaten Error. PM: {}, Climate: {}'.format(resp_pm.reason, resp_bmp.reason))
+            logging.warning(f"Sensor.Community Error. PM: {resp_pm.reason}, Climate: {resp_bmp.reason}")
             return False
     else:
         return False
@@ -186,7 +185,7 @@ def send_to_luftdaten(values, id):
 # Compensation factor for temperature
 comp_factor = 2.25
 
-# Raspberry Pi ID to send to Luftdaten
+# Raspberry Pi ID to send to Sensor.Community
 id = "raspi-" + get_serial_number()
 
 # Width and height to calculate text position
@@ -198,13 +197,14 @@ font_size = 16
 font = ImageFont.truetype(UserFont, font_size)
 
 # Log Raspberry Pi serial and Wi-Fi status
-logging.info("Raspberry Pi serial: {}".format(get_serial_number()))
-logging.info("Wi-Fi: {}\n".format("connected" if check_wifi() else "disconnected"))
+logging.info(f"Raspberry Pi serial: {get_serial_number()}")
+wifi_status = "connected" if check_wifi() else "disconnected"
+logging.info(f"Wi-Fi: {wifi_status}\n")
 
 time_since_update = 0
 update_time = time.time()
 
-# Main loop to read data, display, and send to Luftdaten
+# Main loop to read data, display, and send to Sensor.Community
 while True:
     try:
         values = read_values()
@@ -212,10 +212,10 @@ while True:
         if time_since_update > 145:
             logging.info(values)
             update_time = time.time()
-            if send_to_luftdaten(values, id):
-                logging.info("Luftdaten Response: OK")
+            if send_to_sensorcommunity(values, id):
+                logging.info("Sensor.Community Response: OK")
             else:
-                logging.warning("Luftdaten Response: Failed")
+                logging.warning("Sensor.Community Response: Failed")
         display_status()
     except Exception as e:
-        logging.warning('Main Loop Exception: {}'.format(e))
+        logging.warning(f"Main Loop Exception: {e}")