+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+1.0.1
+-----------
+
+* README.md: Fix images
+
+1.0.0
+-----
+
+* BREAKING: Port to gpiod/gpiodevice for Pi 5/Bookworm.
+
 0.0.6
 -----